Move get_subject_session_list from pipeline/engine.py to utils/io.py
